requirements.txt
python-telegram-bot==20.8
httpx==0.25.2
requests==2.31.0
